qhweb/requirements.txt
altgraph==0.9
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bdist-mpkg==0.4.4
billiard==3.3.0.23
boto==2.45.0
celery==3.1.0
celerymon==1.0.3
certifi==2016.2.28
cffi==1.9.1
cryptography==1.7.2
Cython==0.18
Django==1.6.11
django-admin-enhancer==0.1.2
django-appconf==0.6
django-celery==3.1.17
django-debug-toolbar==0.8.5
django-easy-maps==0.7.4
django-extensions==1.0.2
django-imagekit==2.0.2
django-mobi==0.1.6
django-registration==0.8
django-relatedadminwidget==0.0.3
django-storages==1.1.8
django-storages-redux==1.3
django-wkhtmltopdf==3.0.0
enum34==1.1.6
geopy==0.95
google-api-python-client==1.1
gunicorn==0.17.2
html5lib==0.95
httplib2==0.9.2
idna==2.2
ipaddress==1.0.18
johnny-cache==1.4
kombu==3.0.37
macholib==1.4.2
meld3==1.0.2
mock==1.0.1
modulegraph==0.9.1
MySQL-python==1.2.5
mysqlclient==1.3.7
oauth==1.0.1
oauth2==1.5.211
oauth2client==4.0.0
olefile==0.44
packaging==16.8
Pillow==3.1.0
py-wkhtmltox==0.2
py2app==0.6.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
pyOpenSSL==16.2.0
pyparsing==2.1.10
pyPdf==1.13
PyPDF2==1.25.1
PyRSS2Gen==1.0.0
python-dateutil==2.1
python-gflags==3.0.4
python-memcached==1.48
python-openid==2.2.5
pytz==2016.10
reportlab==2.6
rsa==3.4.2
singledispatch==3.4.0.3
six==1.10.0
South==0.7.6
sqlparse==0.1.18
supervisor==3.2.0
tornado==3.0.1
Twisted==12.0.0
vine==1.1.3
xattr==0.6.2
xhtml2pdf==0.0.6
zope.interface==3.5.1
